In this commit I added a gear system, where the up key and the down key can be used to shift up and down a gear respectively.
This works by calling the shiftUp and shiftDown functions in the car
class, which increase or decrease acceleration based on the values
stored in the settings class, as well as resetting the variables used in
the kinematic equations, namely updating the starting position and the
starting velocity of the new motion.

I have also started work on the system which will allow the car to blow
if the engine is revved too high, by again comparing to values within
the settings file. This isn't implemented yet, however. The next step is
going to be adding the objects to display car speed, gear and time at
the bottom of the screen. After, the reset game functionality will be
implemented so that the game can be played multiple times.